Update Items.js
If we set maxW to useMobileValue, this cause the width of items not to match the select width when we use desktop. useMobileValue only works if we use a mobile device. But if we set the maxW to 100%, than it becomes responsive and the width of select and select items is the same. on all devices.